requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
python-dotenv==1.0.0
